CcTrayViewAuthority => GoConfigPipelinePermissionsAuthority
Also,
* Rename package domain.cctray.viewers to config.security.viewers
