+++ server/src/com/thoughtworks/go/config/security/viewers/NoOne.java
@@ -1,30 +1,30 @@
-/*************************GO-LICENSE-START*********************************
- * Copyright 2015 ThoughtWorks, Inc.
+/*
+ * Copyright 2017 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *************************GO-LICENSE-END***********************************/
+ */
 
-package com.thoughtworks.go.domain.cctray.viewers;
+package com.thoughtworks.go.config.security.viewers;
 
-/* Claims that it has all viewers. */
-public class Everyone implements Viewers {
-    public static Viewers INSTANCE = new Everyone();
+/* Denies that it has any viewers. */
+public class NoOne implements Viewers {
+    public static Viewers INSTANCE = new NoOne();
 
-    private Everyone() {
+    private NoOne() {
     }
 
     @Override
     public boolean contains(String username) {
-        return true;
+        return false;
     }
 }
