+++ server/src/com/thoughtworks/go/domain/cctray/CcTrayConfigChangeHandler.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2015 ThoughtWorks, Inc.
+ * Copyright 2017 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,9 +17,10 @@
 package com.thoughtworks.go.domain.cctray;
 
 import com.thoughtworks.go.config.*;
+import com.thoughtworks.go.config.security.GoConfigPipelinePermissionsAuthority;
 import com.thoughtworks.go.domain.PipelineGroupVisitor;
 import com.thoughtworks.go.domain.activity.ProjectStatus;
-import com.thoughtworks.go.domain.cctray.viewers.Viewers;
+import com.thoughtworks.go.config.security.viewers.Viewers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -32,13 +33,13 @@ import java.util.Map;
 public class CcTrayConfigChangeHandler {
     private CcTrayCache cache;
     private CcTrayStageStatusLoader stageStatusLoader;
-    private CcTrayViewAuthority ccTrayViewAuthority;
+    private GoConfigPipelinePermissionsAuthority pipelinePermissionsAuthority;
 
     @Autowired
-    public CcTrayConfigChangeHandler(CcTrayCache cache, CcTrayStageStatusLoader stageStatusLoader, CcTrayViewAuthority ccTrayViewAuthority) {
+    public CcTrayConfigChangeHandler(CcTrayCache cache, CcTrayStageStatusLoader stageStatusLoader, GoConfigPipelinePermissionsAuthority pipelinePermissionsAuthority) {
         this.cache = cache;
         this.stageStatusLoader = stageStatusLoader;
-        this.ccTrayViewAuthority = ccTrayViewAuthority;
+        this.pipelinePermissionsAuthority = pipelinePermissionsAuthority;
     }
 
     public void call(CruiseConfig config) {
@@ -47,7 +48,7 @@ public class CcTrayConfigChangeHandler {
 
     public void call(PipelineConfig pipelineConfig, String pipelineGroup) {
         ArrayList<ProjectStatus> projectStatuses = new ArrayList<>();
-        final Map<String, Viewers> groupsAndTheirViewers = ccTrayViewAuthority.groupsAndTheirViewers();
+        final Map<String, Viewers> groupsAndTheirViewers = pipelinePermissionsAuthority.groupsAndTheirViewers();
         Viewers usersWithViewPermissionsOfThisGroup = groupsAndTheirViewers.get(pipelineGroup);
         updateProjectStatusForPipeline(usersWithViewPermissionsOfThisGroup, pipelineConfig, projectStatuses);
         cache.putAll(projectStatuses);
@@ -55,7 +56,7 @@ public class CcTrayConfigChangeHandler {
 
     private List<ProjectStatus> findAllProjectStatusesForStagesAndJobsIn(CruiseConfig config) {
         final List<ProjectStatus> projectStatuses = new ArrayList<>();
-        final Map<String, Viewers> groupsAndTheirViewers = ccTrayViewAuthority.groupsAndTheirViewers();
+        final Map<String, Viewers> groupsAndTheirViewers = pipelinePermissionsAuthority.groupsAndTheirViewers();
 
         config.accept(new PipelineGroupVisitor() {
             @Override
