+++ server/test/unit/com/thoughtworks/go/domain/cctray/CcTrayConfigChangeHandlerTest.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2015 ThoughtWorks, Inc.
+ * Copyright 2017 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,9 +16,10 @@
 package com.thoughtworks.go.domain.cctray;
 
 import com.thoughtworks.go.config.*;
+import com.thoughtworks.go.config.security.GoConfigPipelinePermissionsAuthority;
 import com.thoughtworks.go.domain.activity.ProjectStatus;
-import com.thoughtworks.go.domain.cctray.viewers.AllowedViewers;
-import com.thoughtworks.go.domain.cctray.viewers.Viewers;
+import com.thoughtworks.go.config.security.viewers.AllowedViewers;
+import com.thoughtworks.go.config.security.viewers.Viewers;
 import com.thoughtworks.go.helper.GoConfigMother;
 import org.junit.After;
 import org.junit.Before;
@@ -46,7 +47,7 @@ public class CcTrayConfigChangeHandlerTest {
     @Mock
     private CcTrayStageStatusLoader stageStatusLoader;
     @Mock
-    private CcTrayViewAuthority ccTrayViewAuthority;
+    private GoConfigPipelinePermissionsAuthority pipelinePermissionsAuthority;
     @Captor
     ArgumentCaptor<List<ProjectStatus>> statusesCaptor;
 
@@ -58,7 +59,7 @@ public class CcTrayConfigChangeHandlerTest {
     public void setUp() throws Exception {
         initMocks(this);
         goConfigMother = new GoConfigMother();
-        handler = new CcTrayConfigChangeHandler(cache, stageStatusLoader, ccTrayViewAuthority);
+        handler = new CcTrayConfigChangeHandler(cache, stageStatusLoader, pipelinePermissionsAuthority);
         pluginRoleUsersStore = PluginRoleUsersStore.instance();
     }
 
@@ -231,7 +232,7 @@ public class CcTrayConfigChangeHandlerTest {
         pluginRoleUsersStore.assignRole("user4", admin);
 
         AllowedViewers allowedViewers = new AllowedViewers(s("user3"), Collections.singleton(admin));
-        when(ccTrayViewAuthority.groupsAndTheirViewers()).thenReturn(m("group1", viewers("user1", "user2"), "group2", allowedViewers));
+        when(pipelinePermissionsAuthority.groupsAndTheirViewers()).thenReturn(m("group1", viewers("user1", "user2"), "group2", allowedViewers));
 
         CruiseConfig config = GoConfigMother.defaultCruiseConfig();
         goConfigMother.addPipelineWithGroup(config, "group2", "pipeline2", "stage2", "job2");
@@ -310,7 +311,7 @@ public class CcTrayConfigChangeHandlerTest {
         ProjectStatus statusOfPipeline1JobInCache = new ProjectStatus(pipeline1job, "OldActivity-Job", "OldStatus-Job", "OldLabel-Job", new Date(), "p1-job-url");
         when(cache.get(pipeline1Stage)).thenReturn(statusOfPipeline1StageInCache);
         when(cache.get(pipeline1job)).thenReturn(statusOfPipeline1JobInCache);
-        when(ccTrayViewAuthority.groupsAndTheirViewers()).thenReturn(m("group1", viewers("user1", "user2"), "group2", viewers("user3")));
+        when(pipelinePermissionsAuthority.groupsAndTheirViewers()).thenReturn(m("group1", viewers("user1", "user2"), "group2", viewers("user3")));
 
         PipelineConfig pipeline1Config = GoConfigMother.pipelineHavingJob("pipeline1", "stage1", "job1", "arts", "dir").pipelineConfigByName(new CaseInsensitiveString("pipeline1"));
