+++ server/test/unit/com/thoughtworks/go/domain/cctray/CcTrayJobStatusChangeHandlerTest.java
@@ -19,8 +19,8 @@ package com.thoughtworks.go.domain.cctray;
 import com.thoughtworks.go.config.PluginRoleConfig;
 import com.thoughtworks.go.domain.JobInstance;
 import com.thoughtworks.go.domain.activity.ProjectStatus;
-import com.thoughtworks.go.domain.cctray.viewers.AllowedViewers;
-import com.thoughtworks.go.domain.cctray.viewers.Viewers;
+import com.thoughtworks.go.config.security.viewers.AllowedViewers;
+import com.thoughtworks.go.config.security.viewers.Viewers;
 import com.thoughtworks.go.helper.JobInstanceMother;
 import org.junit.Before;
 import org.junit.Test;
